requirements.txt
cookiecutter>=1.6
pytest
gitpython
configparser
